Fix (at least partial) for the 512GB limit bug.  We need to rationalize the off_t vs int64_t vs uint64_t issue.
NOTE: this changes the disk header format which will result in a cache *WIPE*.  This is
unfortunately unavoidable.


git-svn-id: https://svn.apache.org/repos/asf/trafficserver/traffic/trunk@1101758 13f79535-47bb-0310-9956-ffa450edef68

+++ iocore/cache/CacheDisk.cc
@@ -140,7 +140,7 @@ CacheDisk::openStart(int event, void *data)
     return openDone(EVENT_IMMEDIATE, 0);
   }
 
-  if (header->magic != DISK_HEADER_MAGIC || header->num_blocks != len) {
+  if (header->magic != DISK_HEADER_MAGIC || header->num_blocks != (uint64_t)len) {
     Warning("disk header different for disk %s: clearing the disk", path);
     SET_HANDLER(&CacheDisk::clearDone);
     clearDisk();
